pyproject.toml
[tool.poetry]
name = "pycomponents"
version = "0.1.0"
description = "An experimental tool to generate CycloneDX BOM from running Python processes"
authors = ["Manabu Niseki <manabu.niseki@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
cyclonedx-python-lib = "^2.3.0"
loguru = "^0.6.0"
psutil = "^5.9.0"
sh = "^1.14.2"
typer = "^0.4.1"

[tool.poetry.dev-dependencies]
autoflake = "^1.4"
autopep8 = "^1.6.0"
black = "^22.3.0"
coveralls = "^3.3.1"
isort = "^5.10.1"
mypy = "^0.942"
pre-commit = "^2.18.1"
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
pytest-mock = "^3.7.0"
pytest-randomly = "^3.11.0"
pytest-sugar = "^0.9.4"
pyupgrade = "^2.32.0"

[tool.poetry.scripts]
pycomponents = "pycomponents.cli:app"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
